fix tests and extract auth mechanism properties utils
