+++ packages/connect-form/src/hooks/use-connect-form.spec.ts
@@ -1,9 +1,6 @@
 import { expect } from 'chai';
 import ConnectionStringUrl from 'mongodb-connection-string-url';
-import {
-  handleConnectionFormFieldUpdate,
-  parseAuthMechanismProperties,
-} from './use-connect-form';
+import { handleConnectionFormFieldUpdate } from './use-connect-form';
 
 describe('use-connect-form hook', function () {
   describe('#handleConnectionFormFieldUpdate', function () {
@@ -823,28 +820,6 @@ describe('use-connect-form hook', function () {
           ).searchParams.get('authMechanism')
         ).to.equal('PLAIN');
       });
-    });
-  });
-
-  describe('parseAuthMechanismProperties', function () {
-    it('parses legit authMechanismProperties string', function () {
-      const props = parseAuthMechanismProperties(
-        new ConnectionStringUrl(
-          'mongodb://localhost?authMechanismProperties=SERVICE_NAME:serviceName'
-        )
-      );
-
-      expect(props.get('SERVICE_NAME')).to.equal('serviceName');
-    });
-
-    it('parses broken authMechanismProperties string as empty', function () {
-      const props = parseAuthMechanismProperties(
-        new ConnectionStringUrl(
-          'mongodb://localhost?authMechanismProperties=broken'
-        )
-      );
-
-      expect(props.get('SERVICE_NAME')).to.equal(undefined);
     });
   });
 });
